// frontend/CarRental/src/app/models/customer.ts
export class Customer {
    id?: number;
    driverLicenceNumber?: string;
    name?: string;
    firstname?: string;
    adress?: string;
    admin?: boolean; 
    login?: string;
    pwd?: string;
    token?: string;
    birth?: Date;
    vehicleClass?: string; 
    drivingLicenceGetting?: Date; 
    drivingLicenceExpiration?: Date; 
  }